s3_uploader/requirements.txt
streamlit>=1.28.0
boto3>=1.26.0
requests>=2.28.0
python-dotenv>=1.0.0